fix: resolve all 42 Pyright type errors
- settings.py: walrus operator for os.getenv() narrowing
- database.py: property guard for Optional sqlite3.Connection
- types.py: Optional datetime annotations
- streamer.py: correct websockets type, assert guard
- indicators.py: float() wrapper on numpy return
- logger.py: getattr for custom LogRecord attributes
- backtest_interval.py: pd.Index for DataFrame columns

+++ live_trading_bot/data/indicators.py
@@ -28,7 +28,7 @@ class Indicators:
         avg_gain = np.mean(gains)
         avg_loss = np.mean(losses)
         rs = avg_gain / max(avg_loss, 1e-10)
-        return 100 - 100 / (1 + rs)
+        return float(100 - 100 / (1 + rs))
 
     @staticmethod
     def calc_atr(
